% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as.option}
\alias{as.option}
\title{Convert a Result to an Option}
\usage{
as.option(x)
}
\arguments{
\item{x}{the Result}
}
\value{
If Ok(x), returns \code{some(x)}, if Error, returns \code{none}.
}
\description{
Convert a Result to an Option
}
\examples{
as.option(ok(1)) == some(1)
as.option(error()) == none
}
