% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{default_none_predicates}
\alias{default_none_predicates}
\title{Default predicates defining None values.}
\usage{
default_none_predicates()
}
\value{
list(missing, is.null, is.na)
}
\description{
Values that are missing, \code{NULL}, or \code{NA} will be considered \code{None}.
}
\seealso{
missing is.null is.na
}
