% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{error}
\alias{error}
\title{Error type}
\usage{
error(type, message, ...)
}
\arguments{
\item{type}{an error type}

\item{message}{an error message associated with the type}

\item{...}{passed as metadata on the error}
}
\description{
Error type
}
\examples{
e <- error('render_error', 'unable to render Rmd')
\dontrun{
stop(e)
}
}
