% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{expect_err}
\alias{expect_err}
\title{Throws error if Result\if{html}{\out{<Ok>}}}
\usage{
expect_err(x, msg)
}
\arguments{
\item{x}{a Result}

\item{msg}{the error message to throw}
}
\description{
Throws error if Result\if{html}{\out{<Ok>}}
}
\examples{
expect_err(error(), "Should give generic_result_error")
\dontrun{
expect_err(ok(1), "Got OK!, Expected a Result<Error>")
}
}
