% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{match_result}
\alias{match_result}
\title{Pattern match a result output}
\usage{
match_result(x, ...)
}
\arguments{
\item{x}{a Result}

\item{...}{passed to \code{switch}}
}
\description{
Pattern match a result output
}
