% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{!.option}
\alias{!.option}
\title{Syntatic sugar to unwrap an Option}
\usage{
\method{!}{option}(e1)
}
\description{
EXPERIMENTAL & maybe a footgun. Ex if you \code{!x} thinking \code{x} is an option,
but it's an \code{int}.
}
\examples{
!some(5) == 5
\dontrun{
# This will error!
!none
}
}
