% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{ok_or}
\alias{ok_or}
\title{Convert an option to Result<T,E>}
\usage{
ok_or(x, .error_type, ...)
}
\arguments{
\item{x}{an Option}

\item{...}{passed as metadata on the error}
}
\description{
Convert an option to Result<T,E>
}
\examples{
ok_or(some(1), "expected_some")
ok_or(none, "expected_some")
}
