% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{option}
\alias{option}
\title{Coerce value to an option}
\usage{
option(x, .none_predicates = default_none_predicates())
}
\arguments{
\item{x}{a value to convert to \code{some()} or \code{none} based on a set of predicates.}

\item{.none_predicates}{a list of predicate functions that evaluate to
TRUE/FALSE. If a predicate evalutes to \code{TRUE}, we return \code{none}, else we
return \code{some(x)}. Default: \code{default_none_predicates()}}
}
\value{
an option
}
\description{
Coerce value to an option
}
\seealso{
default_none_predicates
}
