% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{try_result}
\alias{try_result}
\title{Force expression to return a Result<T,E>}
\usage{
try_result(expr, ...)
}
\arguments{
\item{expr}{an expression}

\item{...}{passed to \code{error()} if expr fails.}
}
\description{
Use in place of \code{TryCatch}.
}
\examples{
result <- try_result({stop("oh no!")})
result_custom_err <- try_result({stop("oh no!")}, .err_type = "my_custom_error")
result_custom_err$error_type
}
