% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{unwrap.option}
\alias{unwrap.option}
\title{Extract a contained value in an option or error}
\usage{
\method{unwrap}{option}(x)
}
\arguments{
\item{x}{a value. If \code{some()} will extract the contained value. If \code{none}
will crash.}
}
\value{
the unwrapped value in some(x)
}
\description{
Extract a contained value in an option or error
}
