% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{unwrap.result}
\alias{unwrap.result}
\title{Extract the Ok value in a result or error}
\usage{
\method{unwrap}{result}(x)
}
\arguments{
\item{x}{a value. If \code{ok()} will extract the contained value. If \code{error}
will crash.}
}
\value{
the unwrapped value in ok(x)
}
\description{
Extract the Ok value in a result or error
}
