% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{unwrap_err}
\alias{unwrap_err}
\title{Return Error or panic with Ok() value.}
\usage{
unwrap_err(x)
}
\arguments{
\item{x}{a Result}
}
\value{
an Error if Result\if{html}{\out{<Error>}}
}
\description{
Return Error or panic with Ok() value.
}
\examples{
unwrap_err(error())
\dontrun{
unwrap_err(ok(1))
}
}
