% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{unwrap_or}
\alias{unwrap_or}
\title{Unwrap or return a default value}
\usage{
unwrap_or(x, .default)
}
\arguments{
\item{x}{a value}

\item{.default}{the default to return}
}
\description{
Unwrap or return a default value
}
