% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{unwrap_or_else}
\alias{unwrap_or_else}
\title{Unwrap or return a function result}
\usage{
unwrap_or_else(x, .fn)
}
\arguments{
\item{x}{a value}

\item{.fn}{the function to evaluate}
}
\value{
unwrap(x) or the return of \code{.fn()}
}
\description{
Unwrap or return a function result
}
\examples{
unwrap_or_else(none, function() {"whoops!"})
unwrap_or_else(some(5), function() {"whoops!"})
}
